pyproject.toml
[project]
name = "audioconverter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "google-generativeai==0.3.2",
    "openai-whisper>=20250625",
]
